test: add tests
Refs: TSK-92
